reorganization of data checkers
