+++ moirae/estimators/offline/DataCheckers/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tools to help check if data being used is appropriate for offline estimation
+"""
