+++ moirae/estimators/offline/extractors/base.py
@@ -8,30 +8,6 @@ from battdat.data import BatteryDataset
 from moirae.models.base import HealthVariable
 
 
-class DataCheckError(ValueError):
-    """
-    Custom exception to be used when we encounter issues when checking the provided data for features to extract
-    """
-    def __init__(self, message: str = "Data check failed!"):
-        super().__init__(message)
-
-
-class BaseDataChecker():
-    """
-    Base class for tools to check if data being used is appropriate for extractors
-    """
-
-    def check(self, data: Union[pd.DataFrame, BatteryDataset]) -> None:
-        """
-        Verify whether data contains features needed for algorithm
-
-        Args:
-            data: Data to be evaluated
-        Raises:
-            (DataCheckError) If the dataset is missing critical information
-        """
-        pass  # Default: all data is valid
-
 class BaseExtractor:
     """Base class for tools which determine parameters from special cycles"""
