+++ tests/estimators/offline/data_checkers/RPT/test_cap_check.py
@@ -1,0 +1,10 @@
+"""
+Unit test for the ECM data checkers.
+"""
+import pandas as pd
+import numpy as np
+
+from pytest import fixture, raises
+
+from moirae.estimators.offline.DataCheckers.base import DataCheckError
+from moirae.estimators.offline.DataCheckers.RPT import CapacityDataChecker
